refactor(server): decouple SettingsService into a passive data store + coordinator
Introduce SettingsStoreService as the sole owner of persisted settings state
(the hydrated fields, init(), reads, public masking, and low-level
persistence), depending only on the settings repositories. SettingsService
becomes a pure coordinator (test/save/reinit flows) with no settings state of
its own, and every read consumer injects the store.

Because the API services and media-server adapters/factory now read settings
from the store instead of SettingsService, the circular dependencies are gone:
constructor forwardRef injections drop from 28 to 2 and the @swc/jest type-only
import aliases from 28 to 2 -- the only remaining pair is the genuine, mutual
MediaServerFactory <-> MediaServerSwitchService cycle (factory checks
isSwitching(); switch calls uninitializeServer()).

No behaviour change: SettingsService keeps its public API by delegating reads to
the store; save flows persist via the store and re-hydrate. Public settings
masking, save/test/connection, and media-server switch flows are preserved.
