+++ apps/server/src/modules/api/internal-api/internal-api.service.ts
@@ -1,6 +1,5 @@
-import { forwardRef, Inject, Injectable } from '@nestjs/common';
-import { SettingsService } from '../../../modules/settings/settings.service';
-import type { SettingsService as SettingsServiceType } from '../../../modules/settings/settings.service';
+import { Injectable } from '@nestjs/common';
+import { SettingsStoreService } from '../../../modules/settings/settings-store.service';
 import { MaintainerrLogger } from '../../logging/logs.service';
 import { InternalApi } from './helpers/internal-api.helper';
 
@@ -9,8 +8,7 @@ export class InternalApiService {
   private api: InternalApi;
 
   constructor(
-    @Inject(forwardRef(() => SettingsService))
-    private readonly settings: SettingsServiceType,
+    private readonly settings: SettingsStoreService,
     private readonly logger: MaintainerrLogger,
   ) {}
