+++ apps/server/src/modules/api/media-server/plex/plex-adapter.service.ts
@@ -16,12 +16,11 @@ import {
   UpdateCollectionParams,
   WatchRecord,
 } from '@maintainerr/contracts';
-import { forwardRef, Inject, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { MaintainerrLogger } from '../../../logging/logs.service';
 import { EPlexDataType } from '../../plex-api/enums/plex-data-type-enum';
 import { PLEX_PAGE_SIZE } from '../../plex-api/plex-api.constants';
 import { PlexApiService } from '../../plex-api/plex-api.service';
-import type { PlexApiService as PlexApiServiceType } from '../../plex-api/plex-api.service';
 import {
   isBlankMediaServerId,
   isForeignServerId,
@@ -44,8 +43,7 @@ import { PlexMapper } from './plex.mapper';
 @Injectable()
 export class PlexAdapterService implements IMediaServerService {
   constructor(
-    @Inject(forwardRef(() => PlexApiService))
-    private readonly plexApi: PlexApiServiceType,
+    private readonly plexApi: PlexApiService,
     private readonly logger: MaintainerrLogger,
   ) {
     this.logger.setContext(PlexAdapterService.name);
