+++ apps/server/src/modules/api/servarr-api/servarr.service.ts
@@ -1,6 +1,5 @@
-import { forwardRef, Inject, Injectable } from '@nestjs/common';
-import { SettingsService } from '../../../modules/settings/settings.service';
-import type { SettingsService as SettingsServiceType } from '../../../modules/settings/settings.service';
+import { Injectable } from '@nestjs/common';
+import { SettingsStoreService } from '../../../modules/settings/settings-store.service';
 import { MaintainerrLoggerFactory } from '../../logging/logs.service';
 import { RadarrSettingRawDto } from "../../settings/dto's/radarr-setting.dto";
 import { SonarrSettingRawDto } from "../../settings/dto's/sonarr-setting.dto";
@@ -15,8 +14,7 @@ export class ServarrService {
   private sonarrApiCache: Record<string, SonarrApi> = {};
 
   constructor(
-    @Inject(forwardRef(() => SettingsService))
-    private readonly settings: SettingsServiceType,
+    private readonly settings: SettingsStoreService,
     private readonly loggerFactory: MaintainerrLoggerFactory,
   ) {}
