+++ apps/server/src/modules/api/tmdb-api/tmdb.service.ts
@@ -1,9 +1,8 @@
 import { BasicResponseDto, MaintainerrEvent } from '@maintainerr/contracts';
-import { forwardRef, Inject, Injectable } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { OnEvent } from '@nestjs/event-emitter';
 import { MaintainerrLogger } from '../../logging/logs.service';
-import { SettingsService } from '../../settings/settings.service';
-import type { SettingsService as SettingsServiceType } from '../../settings/settings.service';
+import { SettingsStoreService } from '../../settings/settings-store.service';
 import {
   formatConnectionFailureMessage,
   logConnectionTestError,
@@ -22,8 +21,7 @@ const TMDB_DEFAULT_API_KEY = 'db55323b8d3e4154498498a75642b381';
 @Injectable()
 export class TmdbApiService extends ExternalApiService {
   constructor(
-    @Inject(forwardRef(() => SettingsService))
-    private readonly settings: SettingsServiceType,
+    private readonly settings: SettingsStoreService,
     protected readonly logger: MaintainerrLogger,
   ) {
     logger.setContext(TmdbApiService.name);
