+++ apps/server/src/modules/collections/collection-handler.ts
@@ -6,7 +6,7 @@ import { IMediaServerService } from '../api/media-server/media-server.interface'
 import { SeerrApiService } from '../api/seerr-api/seerr-api.service';
 import { MaintainerrLogger } from '../logging/logs.service';
 import { MetadataService } from '../metadata/metadata.service';
-import { SettingsService } from '../settings/settings.service';
+import { SettingsStoreService } from '../settings/settings-store.service';
 import { CollectionsService } from './collections.service';
 import { Collection } from './entities/collection.entities';
 import { CollectionMedia } from './entities/collection_media.entities';
@@ -19,7 +19,7 @@ export class CollectionHandler {
     private readonly mediaServerFactory: MediaServerFactory,
     private readonly collectionService: CollectionsService,
     private readonly seerrApi: SeerrApiService,
-    private readonly settings: SettingsService,
+    private readonly settings: SettingsStoreService,
     private readonly metadataService: MetadataService,
     private readonly radarrActionHandler: RadarrActionHandler,
     private readonly sonarrActionHandler: SonarrActionHandler,
