+++ apps/server/src/modules/notifications/agents/email.ts
@@ -1,7 +1,7 @@
 import type { EmailOptions } from 'email-templates';
 import path from 'path';
 import { MaintainerrLogger } from '../../logging/logs.service';
-import { SettingsService } from '../../settings/settings.service';
+import { SettingsStoreService } from '../../settings/settings-store.service';
 import PreparedEmail from '../email/preparedEmail';
 import { Notification } from '../entities/notification.entities';
 import {
@@ -13,7 +13,7 @@ import type { NotificationAgent, NotificationPayload } from './agent';
 
 class EmailAgent implements NotificationAgent {
   public constructor(
-    private readonly appSettings: SettingsService,
+    private readonly appSettings: SettingsStoreService,
     private readonly settings: NotificationAgentEmail,
     private readonly logger: MaintainerrLogger,
     readonly notification: Notification,
