+++ apps/server/src/modules/notifications/agents/pushbullet.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { MaintainerrLogger } from '../../logging/logs.service';
-import { SettingsService } from '../../settings/settings.service';
+import { SettingsStoreService } from '../../settings/settings-store.service';
 import { Notification } from '../entities/notification.entities';
 import {
   NotificationAgentKey,
@@ -19,7 +19,7 @@ interface PushbulletPayload {
 
 class PushbulletAgent implements NotificationAgent {
   public constructor(
-    private readonly appSettings: SettingsService,
+    private readonly appSettings: SettingsStoreService,
     private readonly settings: NotificationAgentPushbullet,
     private readonly logger: MaintainerrLogger,
     readonly notification: Notification,
