+++ apps/server/src/modules/notifications/agents/pushover.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { MaintainerrLogger } from '../../logging/logs.service';
-import { SettingsService } from '../../settings/settings.service';
+import { SettingsStoreService } from '../../settings/settings-store.service';
 import { Notification } from '../entities/notification.entities';
 import {
   NotificationAgentKey,
@@ -28,7 +28,7 @@ interface PushoverPayload extends PushoverImagePayload {
 
 class PushoverAgent implements NotificationAgent {
   public constructor(
-    private readonly appSettings: SettingsService,
+    private readonly appSettings: SettingsStoreService,
     private readonly settings: NotificationAgentPushover,
     private readonly logger: MaintainerrLogger,
     readonly notification: Notification,
