+++ apps/server/src/modules/notifications/agents/slack.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { MaintainerrLogger } from '../../logging/logs.service';
-import { SettingsService } from '../../settings/settings.service';
+import { SettingsStoreService } from '../../settings/settings-store.service';
 import { Notification } from '../entities/notification.entities';
 import {
   NotificationAgentKey,
@@ -50,7 +50,7 @@ interface SlackBlockEmbed {
 
 class SlackAgent implements NotificationAgent {
   public constructor(
-    private readonly appSettings: SettingsService,
+    private readonly appSettings: SettingsStoreService,
     private readonly settings: NotificationAgentSlack,
     private readonly logger: MaintainerrLogger,
     readonly notification: Notification,
