+++ apps/server/src/modules/notifications/agents/telegram.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { MaintainerrLogger } from '../../logging/logs.service';
-import { SettingsService } from '../../settings/settings.service';
+import { SettingsStoreService } from '../../settings/settings-store.service';
 import { Notification } from '../entities/notification.entities';
 import {
   NotificationAgentKey,
@@ -29,7 +29,7 @@ class TelegramAgent implements NotificationAgent {
   private baseUrl = 'https://api.telegram.org/';
 
   public constructor(
-    private readonly appSettings: SettingsService,
+    private readonly appSettings: SettingsStoreService,
     private readonly settings: NotificationAgentTelegram,
     private readonly logger: MaintainerrLogger,
     readonly notification: Notification,
