+++ apps/server/src/modules/rules/tasks/rule-executor.service.ts
@@ -26,7 +26,7 @@ import {
   RuleHandlerFailedDto,
 } from '../../events/events.dto';
 import { MaintainerrLogger } from '../../logging/logs.service';
-import { SettingsService } from '../../settings/settings.service';
+import { SettingsStoreService } from '../../settings/settings-store.service';
 import { RuleConstants } from '../constants/rules.constants';
 import { RulesDto } from '../dtos/rules.dto';
 import { RuleGroup } from '../entities/rule-group.entities';
@@ -100,7 +100,7 @@ export class RuleExecutorService {
     private readonly rulesService: RulesService,
     private readonly mediaServerFactory: MediaServerFactory,
     private readonly collectionService: CollectionsService,
-    private readonly settings: SettingsService,
+    private readonly settings: SettingsStoreService,
     private readonly comparatorFactory: RuleComparatorServiceFactory,
     private readonly eventEmitter: EventEmitter2,
     private readonly progressManager: RuleExecutorProgressService,
